Fix super weird ECJ compilation error return null x.binding randomly
Seems like the ReferenceExpression x should be created in some other thread and so the x.binding is still null when the code executes the x.binding call or something like that because it is super difficult to reproduce the problem, but it happens sometimes. Also in small project do not happens, but maybe this is also cause by the tread conditions where in small project might not occurs.

Change-Id: If48cf5a1d9fa053ed9ac186ddafbd0db776a020e